:recycle: クライアントからのメッセージに反応する処理をリファクタリングした
